+++ src/cmd/real_cugan.rs
@@ -1,7 +1,6 @@
 use std::{
     env,
     path::{Path, PathBuf},
-    process::Stdio,
     sync::Arc,
 };
 
@@ -11,14 +10,13 @@ use clap::{value_parser, Args};
 use fluent_templates::Loader;
 use image::io::Reader;
 use novel_api::Timing;
-use parking_lot::RwLock;
 use tokio::{
     fs::File,
     io::{AsyncReadExt, BufReader},
-    process::{Child, Command},
+    process::Command,
     sync::Semaphore,
 };
-use tracing::info;
+use tracing::{error, info};
 use walkdir::WalkDir;
 
 use crate::{
@@ -33,7 +31,7 @@ pub struct RealCugan {
     #[arg(help = LOCALES.lookup(&LANG_ID, "image_path").unwrap())]
     pub image_path: Option<PathBuf>,
 
-    #[arg(short, long, default_value_t = 4, value_parser = value_parser!(u8).range(1..=16),
+    #[arg(short, long, default_value_t = 4, value_parser = value_parser!(u8).range(1..=8),
         help = LOCALES.lookup(&LANG_ID, "maximum_concurrency").unwrap())]
     pub maximum_concurrency: u8,
 }
@@ -48,7 +46,7 @@ pub async fn execute(config: RealCugan) -> Result<()> {
 
     let semaphore = Arc::new(Semaphore::new(config.maximum_concurrency as usize));
     let image_paths = image_paths(config).await?;
-    let pb = Arc::new(RwLock::new(ProgressBar::new(image_paths.len())));
+    let pb = ProgressBar::new(image_paths.len() as u64)?;
 
     let curr_path = env::current_dir()?;
 
@@ -56,7 +54,13 @@ pub async fn execute(config: RealCugan) -> Result<()> {
         let image = Reader::open(&input_path)?.decode()?;
         let scale = calc_scale(image.width(), image.height());
 
-        let output_path = input_path.with_extension(utils::image_ext(&image));
+        let ext = utils::image_ext(&image);
+        if ext.is_err() {
+            error!("{}: {}", ext.unwrap_err(), input_path.display());
+            continue;
+        }
+
+        let output_path = input_path.with_extension(ext.unwrap());
 
         if input_path != output_path {
             to_delete.push(input_path.clone());
@@ -71,14 +75,14 @@ pub async fn execute(config: RealCugan) -> Result<()> {
         );
 
         let permit = semaphore.clone().acquire_owned().await.unwrap();
-        let pb = Arc::clone(&pb);
+        let mut pb = pb.clone();
 
         let absolute_path = dunce::canonicalize(&input_path)?;
         let diff_path = pathdiff::diff_paths(absolute_path, &curr_path).unwrap();
 
         handles.push(tokio::spawn(async move {
-            pb.write().inc(diff_path.display().to_string());
-            create_child(input_path, output_path, scale)?.wait().await?;
+            pb.inc(diff_path.display().to_string());
+            create_child(input_path, output_path, scale).await?;
 
             drop(permit);
 
@@ -90,13 +94,13 @@ pub async fn execute(config: RealCugan) -> Result<()> {
         handle.await??;
     }
 
-    pb.write().finish();
+    pb.finish();
 
     for path in to_delete {
         utils::remove_file_or_dir(path)?;
     }
 
-    info!("Time spent on `info`: {}", timing.elapsed()?);
+    info!("Time spent on `real_cugan`: {}", timing.elapsed()?);
 
     Ok(())
 }
@@ -137,7 +141,7 @@ where
 
     let file = File::open(path).await?;
     let mut reader = BufReader::new(file);
-    let mut buffer = BytesMut::with_capacity(1024);
+    let mut buffer = BytesMut::with_capacity(128);
 
     reader.read_buf(&mut buffer).await?;
 
@@ -163,21 +167,27 @@ const fn calc_scale(width: u32, height: u32) -> u8 {
     }
 }
 
-fn create_child<T, E>(input_path: T, output_path: E, scale: u8) -> Result<Child>
+async fn create_child<T, E>(input_path: T, output_path: E, scale: u8) -> Result<()>
 where
     T: AsRef<Path>,
     E: AsRef<Path>,
 {
-    let child = Command::new("realcugan-ncnn-vulkan")
+    let output = Command::new("realcugan-ncnn-vulkan")
         .arg("-i")
         .arg(input_path.as_ref())
         .arg("-o")
         .arg(output_path.as_ref())
         .arg("-s")
         .arg(scale.to_string())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()?;
+        .output()
+        .await?;
 
-    Ok(child)
+    info!("{}", simdutf8::basic::from_utf8(&output.stdout)?);
+
+    if !output.status.success() {
+        error!("{}", simdutf8::basic::from_utf8(&output.stderr)?);
+        bail!("`realcugan-ncnn-vulkan` failed to execute");
+    }
+
+    Ok(())
 }