+++ src/utils/login.rs
@@ -2,7 +2,7 @@ use std::io::{self, Write};
 
 use anyhow::Result;
 use novel_api::{Client, Keyring, UserInfo};
-use tracing::{info, warn};
+use tracing::info;
 
 use crate::cmd::Source;
 
@@ -27,7 +27,7 @@ where
 
                 client.login(username, password.unwrap()).await?;
             } else {
-                warn!("Unable to get password from Keyring");
+                info!("Unable to get password from Keyring");
 
                 let password = get_password()?;
                 client.login(username, &password).await?;
@@ -58,7 +58,6 @@ fn get_username() -> Result<String> {
     Ok(buffer.to_string())
 }
 
-// TODO i18n output
 fn get_password() -> Result<String> {
     let password = rpassword::prompt_password("Password: ")?;
     Ok(password)