+++ src/utils/progress.rs
@@ -1,15 +1,21 @@
 use std::fmt::Write;
 
+use anyhow::Result;
+use crossterm::terminal;
 use indicatif::{ProgressState, ProgressStyle};
 
 #[must_use]
+#[derive(Clone)]
 pub struct ProgressBar {
     pb: indicatif::ProgressBar,
+    message_width: usize,
 }
 
 impl ProgressBar {
-    pub fn new(total_size: usize) -> Self {
-        let pb = indicatif::ProgressBar::new(total_size as u64);
+    pub fn new(total_size: u64) -> Result<Self> {
+        let message_width = (terminal::size()?.0 / 3) as usize;
+
+        let pb = indicatif::ProgressBar::new(total_size);
 
         pb.set_style(ProgressStyle::with_template("{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos:>7}/{len:7} {msg:40} ({eta})")
         .unwrap()
@@ -18,7 +24,7 @@ impl ProgressBar {
         })
         .progress_chars("#>-"));
 
-        Self { pb }
+        Ok(Self { pb, message_width })
     }
 
     pub fn inc<T>(&mut self, msg: T)
@@ -26,13 +32,12 @@ impl ProgressBar {
         T: AsRef<str>,
     {
         self.pb.set_message(
-            console::truncate_str(msg.as_ref(), (viuer::terminal_size().0 / 3) as usize, "...")
-                .to_string(),
+            console::truncate_str(msg.as_ref(), self.message_width, "...").to_string(),
         );
         self.pb.inc(1);
     }
 
-    pub fn finish(&mut self) {
+    pub fn finish(self) {
         self.pb.finish_and_clear();
     }
 }